LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/html/semantics/forms/constraints/form-validation-validity-valueMissing-expected.txt


PASS [INPUT in TEXT status] The required attribute is not set 
PASS [INPUT in TEXT status] The value is not empty and required is true 
FAIL [INPUT in TEXT status] The value is empty and required is true assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
PASS [INPUT in SEARCH status] The required attribute is not set 
PASS [INPUT in SEARCH status] The value is not empty and required is true 
FAIL [INPUT in SEARCH status] The value is empty and required is true assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
PASS [INPUT in TEL status] The required attribute is not set 
PASS [INPUT in TEL status] The value is not empty and required is true 
FAIL [INPUT in TEL status] The value is empty and required is true assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
PASS [INPUT in URL status] The required attribute is not set 
PASS [INPUT in URL status] The value is not empty and required is true 
FAIL [INPUT in URL status] The value is empty and required is true assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
PASS [INPUT in EMAIL status] The required attribute is not set 
PASS [INPUT in EMAIL status] The value is not empty and required is true 
FAIL [INPUT in EMAIL status] The value is empty and required is true assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
PASS [INPUT in PASSWORD status] The required attribute is not set 
PASS [INPUT in PASSWORD status] The value is not empty and required is true 
FAIL [INPUT in PASSWORD status] The value is empty and required is true assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
PASS [INPUT in DATETIME-LOCAL status] The required attribute is not set 
PASS [INPUT in DATETIME-LOCAL status] Valid local date and time string(2000-12-10T12:00:00) 
FAIL [INPUT in DATETIME-LOCAL status] Valid local date and time string(2000-12-10 12:00) assert_false: The validity.valueMissing should be false. expected false got true
PASS [INPUT in DATETIME-LOCAL status] Valid local date and time string(1979-10-14T12:00:00.001) 
FAIL [INPUT in DATETIME-LOCAL status] The value attribute is a number(1234567) assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in DATETIME-LOCAL status] The value attribute is a Date object assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in DATETIME-LOCAL status] Invalid local date and time string(1979-10-99 99:99) assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in DATETIME-LOCAL status] Valid local date and time string(1979-10-14 12:00:00) assert_false: The validity.valueMissing should be false. expected false got true
FAIL [INPUT in DATETIME-LOCAL status] Invalid local date and time string(2001-12-21  12:00)-two white space assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in DATETIME-LOCAL status] the value attribute is a string(abc) assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in DATETIME-LOCAL status] The value attribute is empty string assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
PASS [INPUT in DATE status] The required attribute is not set 
PASS [INPUT in DATE status] Valid date string(2000-12-10) 
PASS [INPUT in DATE status] Valid date string(9999-01-01) 
FAIL [INPUT in DATE status] The value attribute is a number(1234567) assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in DATE status] The value attribute is a Date object assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in DATE status] Invalid date string(9999-99-99) assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in DATE status] Invalid date string(37-01-01) assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in DATE status] Invalid date string(2000/01/01) assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in DATE status] The value attribute is empty string assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
PASS [INPUT in MONTH status] The required attribute is not set 
PASS [INPUT in MONTH status] Valid month string(2000-12) 
PASS [INPUT in MONTH status] Valid month string(9999-01) 
FAIL [INPUT in MONTH status] The value attribute is a number(1234567) assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in MONTH status] The value attribute is a Date object assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in MONTH status] Invalid month string(2000-99) assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in MONTH status] Invalid month string(37-01) assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in MONTH status] Invalid month string(2000/01) assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in MONTH status] The value attribute is empty string assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
PASS [INPUT in WEEK status] The required attribute is not set 
PASS [INPUT in WEEK status] Valid week string(2000-W12) 
PASS [INPUT in WEEK status] Valid week string(9999-W01) 
FAIL [INPUT in WEEK status] The value attribute is a number(1234567) assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in WEEK status] The value attribute is a Date object assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in WEEK status] Invalid week string(2000-W99) assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in WEEK status] invalid week string(2000-W00) assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in WEEK status] invalid week string(2000-w01) assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in WEEK status] The value attribute is empty string assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
PASS [INPUT in TIME status] The required attribute is not set 
PASS [INPUT in TIME status] Validtime string(12:00:00) 
PASS [INPUT in TIME status] Validtime string(12:00) 
PASS [INPUT in TIME status] Valid time string(12:00:60.001) 
PASS [INPUT in TIME status] Valid time string(12:00:60.01) 
PASS [INPUT in TIME status] Valid time string(12:00:60.1) 
FAIL [INPUT in TIME status] The value attribute is a number(1234567) assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in TIME status] The value attribute is a time object assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in TIME status] Invalid time string(25:00:00) assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in TIME status] Invalid time string(12:60:00) assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in TIME status] Invalid time string(12:00:60) assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in TIME status] Invalid time string(12:00:00:001) assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in TIME status] The value attribute is empty string assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
PASS [INPUT in NUMBER status] The required attribute is not set 
PASS [INPUT in NUMBER status] Value is an integer with a leading symbol '+' 
PASS [INPUT in NUMBER status] Value is a number with a '-' symbol 
PASS [INPUT in NUMBER status] Value is a number in scientific notation form(e is in lowercase) 
PASS [INPUT in NUMBER status] Value is a number in scientific notation form(E is in uppercase) 
PASS [INPUT in NUMBER status] Value is -0 
FAIL [INPUT in NUMBER status] Value is a number with some white spaces assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in NUMBER status] Value is Math.pow(2, 1024) assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in NUMBER status] Value is Math.pow(-2, 1024) assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in NUMBER status] Value is a string that cannot be converted to a number assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
FAIL [INPUT in NUMBER status] The value attribute is empty string assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
PASS [INPUT in CHECKBOX status] The required attribute is not set 
PASS [INPUT in CHECKBOX status] The checked attribute is true 
FAIL [INPUT in CHECKBOX status] The checked attribute is false assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
PASS [INPUT in RADIO status] The required attribute is not set 
PASS [INPUT in RADIO status] The checked attribute is true 
FAIL [INPUT in RADIO status] The checked attribute is false assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
PASS [INPUT in FILE status] The required attribute is not set 
FAIL [INPUT in FILE status] The Files attribute is null assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
PASS [select]  The required attribute is not set 
FAIL [select]  Selected the option with value equals to 1 assert_false: The validity.valueMissing should be false, when control is readonly. expected false got true
FAIL [select]  Selected the option with value equals to empty assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
PASS [textarea]  The required attribute is not set 
PASS [textarea]  The value is not empty 
FAIL [textarea]  The value is empty assert_true: The validity.valueMissing should be true, when control is disabled. expected true got false
